+++ src/index.js
@@ -13,7 +13,7 @@ import Profile from "views/pages/Profile.jsx";
 import Register from "views/pages/Register.jsx";
 import Prediction from "views/dagdi/Pages/Prediction.jsx";
 import PredictionRequirement from "views/dagdi/Pages/PredicitonRequirement.jsx";
-import CorpsProfile from "./views/dagdi/Pages/CorpsProfile";
+import CropsProfile from "./views/dagdi/Pages/CropsProfile";
 import i18n from "./i18n";
 
 ReactDOM.render(
@@ -44,9 +44,9 @@ ReactDOM.render(
                     render={props => <Profile {...props} />}
                 />
                 <Route
-                    path="/corps-page/:id"
+                    path="/crops-page/:id"
                     exact
-                    render={props => <CorpsProfile {...props} />}
+                    render={props => <CropsProfile {...props} />}
                 />
                 <Route
                     path="/register-page"