+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## [1.0.0] 2019-05-13
 ### Original Release
-- Added Base website for (DCPS) dagdi corp prediction system
+- Added Base website for (DCPS) dagdi crop prediction system